Convert portal theme store to TS
